Split admin metrics shared helpers
